# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sheriff Department Shift Scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Deactivate user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SHIFTS
...
  /shifts/{id}:
    get:
      operationId: getShift
      tags: [shifts]
      summary: Get shift by ID with roster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shift with roster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftWithRoster"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateShift
      tags: [shifts]
      summary: Update shift (e.g. change sergeant)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShiftBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SHIFT ASSIGNMENTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        shiftId:
          type: ["integer", "null"]
        shiftName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, sergeant, deputy]
        shiftId:
          type: ["integer", "null"]
      required:
        - email
        - password
        - firstName
        - lastName
        - role
    UpdateUserBody:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, sergeant, deputy]
        shiftId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        password:
          type: string
    Shift:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shiftType:
          type: string
          enum: [day, night]
        shiftLetter:
          type: string
          enum: [a, b]
        sergeantId:
          type: ["integer", "null"]
        sergeantName:
          type: ["string", "null"]
      required:
        - id
        - name
        - shiftType
        - shiftLetter
    ShiftWithRoster:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shiftType:
          type: string
          enum: [day, night]
        shiftLetter:
          type: string
          enum: [a, b]
        sergeantId:
          type: ["integer", "null"]
        sergeantName:
          type: ["string", "null"]
        members:
          type: array
          items:
            $ref: "#/components/schemas/ShiftMember"
      required:
        - id
        - name
        - shiftType
        - shiftLetter
        - members
    ShiftMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        email:
          type: string
      required:
        - id
        - userId
        - firstName
        - lastName
        - role
        - email
    UpdateShiftBody:
      type: object
      properties:
        sergeantId:
          type: ["integer", "null"]
...