# lib/api-zod/src/generated/types/shiftWithRoster.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sheriff Department Shift Scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { ShiftMember } from "./shiftMember";
import type { ShiftWithRosterShiftLetter } from "./shiftWithRosterShiftLetter";
import type { ShiftWithRosterShiftType } from "./shiftWithRosterShiftType";

export interface ShiftWithRoster {
  id: number;
  name: string;
  shiftType: ShiftWithRosterShiftType;
  shiftLetter: ShiftWithRosterShiftLetter;
  /** @nullable */
  sergeantId?: number | null;
  /** @nullable */
  sergeantName?: string | null;
  members: ShiftMember[];
}