# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sheriff Department Shift Scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: string;
  /** @nullable */
  shiftId?: number | null;
  /** @nullable */
  shiftName?: string | null;
  isActive: boolean;
  createdAt: string;
}